+++ crates/forge/tests/cli/main.rs
@@ -20,5 +20,6 @@ mod soldeer;
 mod svm;
 mod test_cmd;
 mod verify;
+mod verify_bytecode;
 
 mod ext_integration;
